+++ project/Grpc.scala
@@ -101,6 +101,13 @@ object Grpc extends Base {
       publishArtifact := false
     )
 
-  val all = aggregateDir("grpc", eg, gen, runtime)
+  val usage = projectDir("grpc/usage")
+    .withGrpc
+    .withTests()
+    .settings(
+      publishArtifact := true
+    )
+
+  val all = aggregateDir("grpc", usage, eg, gen, runtime)
 
 }